// src/presentation/http/dtos/post/posts.request.dto.ts
import { Media } from '@/domain/value-objects/media.value-object';
import { EPostAudience, EPostType } from '@/domain/enums/posts.enum';

import { CursorPaginationQueryDTO } from '@/presentation/http/dtos/common/common.request.dto';

import { ParamsDictionary } from 'express-serve-static-core';

export class CreatePostRequestDTO {
  type: EPostType;
  audience: EPostAudience;
  allowStrangerComments: boolean;
  content: string;
  parentId: string | null;
  hashtags: string[];
  mentions: string[];
  media: Media[];
  constructor(payload: {
    type: EPostType;
    audience: EPostAudience;
    allowStrangerComments: boolean;
    content: string;
    parentId: string | null;
    hashtags: string[];
    mentions: string[];
    media: Media[];
  }) {
    this.type = payload.type;
    this.audience = payload.audience;
    this.allowStrangerComments = payload.allowStrangerComments;
    this.content = payload.content;
    this.parentId = payload.parentId;
    this.hashtags = payload.hashtags;
    this.mentions = payload.mentions;
    this.media = payload.media.map((m) => Media.create({ url: m.url, type: m.type }));
  }
}

export class PatchPostRequestDTO {
  audience: EPostAudience;
  allowStrangerComments: boolean;
  constructor(payload: { audience: EPostAudience; allowStrangerComments: boolean }) {
    this.audience = payload.audience;
    this.allowStrangerComments = payload.allowStrangerComments;
  }
}

export interface GetPostDetailParamsDTO extends ParamsDictionary {
  postId: string;
}

export interface GetPostsParamsDTO extends ParamsDictionary {
  postId: string;
  type: EPostType;
}

export interface GetNewFeedsPayloadDTO {
  userId: string;
  cursor?: CursorPaginationQueryDTO['cursor'];
  limit: CursorPaginationQueryDTO['limit'];
}